pyproject.toml
[project]
name = "multi-doc-chat-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt==0.9",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.12.0",
    "fastapi==0.115.6",
    "ipykernel==6.30.0",
    "jinja2==3.1.4",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-google-genai==2.1.8",
    "langchain-groq==0.3.6",
    "langchain-openai==0.2.10",
    "langsmith>=0.4.41",
    "openai>=1.109.1",
    "pandas>=2.3.3",
    "pydantic>=2.0.0",
    "pypdf>=3.0.0",
    "pyyaml>=6.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "structlog>=25.5.0",
    "uvicorn==0.32.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
